docshell/test/mochitest/test_history_length_during_pageload.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test history.length during page load</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
  <script>
    SimpleTest.waitForExplicitFinish();

    function runTest(testAsyncLength) {
      return new Promise(function(resolve) {
        var initialLength;
        var messageCount = 0;
        var win;
        window.onmessage = function(event) {
          if (event.data.initialLength) {
            initialLength = event.data.initialLength;
          } else {
            is(event.data.length, initialLength + 1,
               "history.length should increase by one when loading a new page");
          }
          if (++messageCount == 3) {
            win.close();
            resolve();
          }
        }
        win = window.open("file_history_length_during_pageload.html?testAsyncLength=" + testAsyncLength);
      });
    }

    async function runTests() {
      await runTest(!SpecialPowers.getBoolPref("fission.sessionHistoryInParent"));
      SimpleTest.finish();
    }
  </script>
</head>
<body onload="runTests();">
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test"></pre>
</body>
</html>
